+++ projects/packages/forms/src/dashboard/admin-migrate-page/index.tsx
@@ -109,10 +109,10 @@ const AdminMigratePage = () => {
 					<Container>
 						<Col lg={ 12 } md={ 8 } sm={ 4 }>
 							<h1 style={ { fontSize: '2.5em', marginBottom: '0.5em' } }>
-								{ __( 'Forms moved', 'jetpack-forms' ) }
+								{ __( 'Forms responses have moved', 'jetpack-forms' ) }
 							</h1>
 							<p style={ { fontSize: '1.7em', marginTop: '0.5em' } }>
-								{ __( "Now it's part of Jetpack → Forms", 'jetpack-forms' ) }
+								{ __( 'They can now be found at Jetpack → Forms', 'jetpack-forms' ) }
 							</p>
 							<p>
 								<Button variant="primary" onClick={ onCheckNewFormsClick }>