+++ projects/js-packages/components/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ### This is a list detailing changes for the Jetpack RNA Components package releases.
 
+## [0.73.4] - 2025-06-02
+### Added
+- Added className prop to AdminPage component. [#43592]
+
+### Changed
+- sass: `@use` rather than `@import` for `@automattic/jetpack-base-styles/gutenberg-base-styles`. [#43607]
+
 ## [0.73.3] - 2025-05-26
 ### Changed
 - Update package dependencies. [#43578]
@@ -1404,6 +1411,7 @@
 ### Changed
 - Update node version requirement to 14.16.1
 
+[0.73.4]: https://github.com/Automattic/jetpack-components/compare/0.73.3...0.73.4
 [0.73.3]: https://github.com/Automattic/jetpack-components/compare/0.73.2...0.73.3
 [0.73.2]: https://github.com/Automattic/jetpack-components/compare/0.73.1...0.73.2
 [0.73.1]: https://github.com/Automattic/jetpack-components/compare/0.73.0...0.73.1